refactor:useBGM과 유틸 분리
